requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
gunicorn==20.1.0
importlib-metadata==6.0.0
Markdown==3.4.1
mysqlclient==2.1.1
psycopg2-binary==2.9.5
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.13.0
